legacy/requirements.txt
streamlit==1.28.2
streamlit-folium==0.16.0
streamlit-authenticator==0.2.3
openai==1.3.5
pyyaml==6.0.1
pytest-cov==4.0.0
flake8==6.0.0
geopy==2.4.1